+++ extensions/vscode/config_schema.json
@@ -3211,14 +3211,6 @@
             "modelRoles": {
               "type": "object",
               "properties": {
-                "inlineEdit": {
-                  "description": "The 'title' property of the model you wish to use for inline edits",
-                  "type": "string"
-                },
-                "applyCodeBlock": {
-                  "description": "The 'title' property of the model you wish to use for applying code blocks",
-                  "type": "string"
-                },
                 "repoMapFileSelection": {
                   "description": "The 'title' property of the model you wish to use for repo map file selections",
                   "type": "string"
